--- apps/documentation/src/pages/atomic/user-message.mdx
@@ -1,17 +1,0 @@
-import { StoryEmbed } from '../../components/StoryEmbed';
-import { ClientRenderer } from '../../components/ClientRenderer';
-
-<div style={{maxWidth: '300px'}}>
-# User Message
-
-## Short Message
-
-<br />
-<StoryEmbed for="UserMessage" name="Short" />
-
-## Long Message
-
-<br />
-<StoryEmbed for="UserMessage" name="Long" />
-
-</div>
