+++ packages/chat/src/components/AgentMessage/index.tsx
@@ -9,9 +9,7 @@ import remarkGfm from 'remark-gfm';
 
 import { FeedbackButton } from '../FeedbackButton';
 import type { IFeedbackButton } from '../FeedbackButton/FeedbackButton.interface';
-import { FeedbackButtonVariant } from '../FeedbackButton/FeedbackButton.interface';
 import { Icon } from '../Icon';
-import { feedbackContainer } from '../SystemResponse/styles.css';
 import {
   aiIconModifier,
   codeBlockContainer,
@@ -44,8 +42,6 @@ export const AgentMessage: React.FC<IAgentMessage> = ({
   text,
   children,
   ai,
-  feedback,
-  isLast,
   generatedMessage = 'Generated by AI, double-check for accuracy.',
 }) => {
   const content = typeof text === 'string' ? text : serializeToMarkdown(text);
@@ -90,14 +86,6 @@ export const AgentMessage: React.FC<IAgentMessage> = ({
         <div className={generatedChin}>
           <Icon svg="ai" className={aiIconModifier} />
           <span>{generatedMessage}</span>
-        </div>
-      )}
-      {feedback && (
-        <div className={feedbackContainer({ isLast })}>
-          <FeedbackButton
-            {...feedback}
-            variant={isLast ? FeedbackButtonVariant.LAST_RESPONSE : FeedbackButtonVariant.PREVIOUS_RESPONSE}
-          />
         </div>
       )}
     </div>
