+++ packages/chat/src/components/MessageContainer/styles.css.ts
@@ -1,0 +1,5 @@
+import { style } from '@vanilla-extract/css';
+
+export const messageContainer = style({
+  margin: '16px 0',
+});
