+++ packages/chat/src/components/UserResponse/index.tsx
@@ -1,11 +1,9 @@
-import clsx from 'clsx';
-
 import { ClassName } from '@/constants';
 import { useAutoScroll } from '@/hooks/useAutoScroll';
 
+import { MessageContainer } from '../MessageContainer';
 import { Tooltip } from '../Tooltip';
-import { UserMessage } from '../UserMessage';
-import { debugMessage, messageContainer, messageRow } from './styles.css';
+import { debugMessage, messageContainer } from './styles.css';
 
 export interface DebugActionProps {
   label: string;
@@ -55,13 +53,13 @@ export const UserResponse: React.FC<UserResponseProps> = ({ message, debug, isFi
   useAutoScroll();
 
   // TODO: Check this in different render modes
+  //       In the 'old' chat app - Messages have a maximum width in different screen modes.
+  //       Should we preserve this ?
   // const { config } = useContext(RuntimeStateAPIContext);
 
   return (
-    <div className={clsx(ClassName.USER_RESPONSE, messageContainer)}>
-      <div className={messageRow({ isFirst, hasAvatar })}>
-        <UserMessage message={message} />
-      </div>
+    <MessageContainer className={ClassName.USER_RESPONSE}>
+      <div className={messageContainer}>{message}</div>
       {debug && (
         <>
           <aside className={debugMessage}>{debug.message}</aside>
@@ -72,6 +70,6 @@ export const UserResponse: React.FC<UserResponseProps> = ({ message, debug, isFi
           )}
         </>
       )}
-    </div>
+    </MessageContainer>
   );
 };
