+++ packages/chat/src/stories.ts
@@ -14,5 +14,4 @@ export * as NewChat from './components/NewChat/NewChat.story';
 export * as Prompt from './components/Prompt/Prompt.story';
 export * as SendButton from './components/SendButton/SendButton.story';
 export * as TypingIndicator from './components/TypingIndicator/TypingIndicator.story';
-export * as UserMessage from './components/UserMessage/UserMessage.story';
 export * as WelcomeMessage from './components/WelcomeMessage/WelcomeMessage.story';
